+++ src/main/java/woowacourse/shoppingcart/dto/request/PasswordRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.request;
 
 public class PasswordRequest {
