+++ src/main/java/woowacourse/shoppingcart/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.response;
 
 public class TokenResponse {
     private String accessToken;
